+++ src/client/java/xyz/kohara/stellarity/StellarityClient.java
@@ -4,10 +4,7 @@ import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import xyz.kohara.stellarity.Stellarity;
-import xyz.kohara.stellarity.registry.StellarityClientNetworking;
-import xyz.kohara.stellarity.registry.StellarityEntityRenderers;
-import xyz.kohara.stellarity.registry.StellarityModels;
-import xyz.kohara.stellarity.registry.StellarityTooltips;
+import xyz.kohara.stellarity.registry.*;
 
 @Environment(EnvType.CLIENT)
 public class StellarityClient implements ClientModInitializer {
@@ -16,6 +13,7 @@ public class StellarityClient implements ClientModInitializer {
         Stellarity.LOGGER.info("Stellarity Client Initializing");
 
         StellarityModels.init();
+        StellarityClientParticles.init();
         StellarityEntityRenderers.init();
         StellarityTooltips.init();
         StellarityClientNetworking.init();